+++ src/tests/EntityGraphQL.Tests/GraphQLSchemaGenerateTests.cs
@@ -16,31 +16,31 @@ namespace EntityGraphQL.Tests
         [Fact]
         public void TestIgnoreQueryFails()
         {
-            var schemaProvider = SchemaBuilder.FromObject<IgnoreTestSchema>(false);
+            var schemaProvider = SchemaBuilder.FromObject<IgnoreTestSchema, object>(false);
             // Add a argument field with a require parameter
             var gql = new QueryRequest {
                 Query = @"query Test { movies { id } }",
             };
-            dynamic results = new IgnoreTestSchema().QueryObject(gql, schemaProvider).Errors;
+            dynamic results = schemaProvider.ExecuteQuery(gql, new IgnoreTestSchema(), null, null).Errors;
             var err = Enumerable.First(results);
             Assert.Equal("Field 'movies' not found on current context 'IgnoreTestSchema'", err.Message);
         }
         [Fact]
         public void TestIgnoreQueryPasses()
         {
-            var schemaProvider = SchemaBuilder.FromObject<IgnoreTestSchema>(false);
+            var schemaProvider = SchemaBuilder.FromObject<IgnoreTestSchema, object>(false);
             // Add a argument field with a require parameter
             var gql = new QueryRequest {
                 Query = @"query Test { albums { id } }",
             };
-            var results = new IgnoreTestSchema().QueryObject(gql, schemaProvider);
+            var results = schemaProvider.ExecuteQuery(gql, new IgnoreTestSchema(), null, null);
             Assert.Empty(((IEnumerable)results.Data["albums"]));
         }
 
         [Fact]
         public void TestIgnoreInputFails()
         {
-            var schemaProvider = SchemaBuilder.FromObject<IgnoreTestSchema>(false);
+            var schemaProvider = SchemaBuilder.FromObject<IgnoreTestSchema, object>(false);
             schemaProvider.AddMutationFrom(new IgnoreTestMutations());
             // Add a argument field with a require parameter
             var gql = new QueryRequest {
@@ -54,7 +54,7 @@ namespace EntityGraphQL.Tests
                     {"hiddenInputField", "yeh"},
                 }
             };
-            var results = new IgnoreTestSchema().QueryObject(gql, schemaProvider);
+            var results = schemaProvider.ExecuteQuery(gql, new IgnoreTestSchema(), null, null);
             var error = results.Errors.First();
             Assert.Equal("No argument 'hiddenInputField' found on field 'addAlbum'", error.Message);
         }
@@ -62,7 +62,7 @@ namespace EntityGraphQL.Tests
         [Fact]
         public void TestIgnoreInputPasses()
         {
-            var schemaProvider = SchemaBuilder.FromObject<IgnoreTestSchema>(false);
+            var schemaProvider = SchemaBuilder.FromObject<IgnoreTestSchema, object>(false);
             schemaProvider.AddMutationFrom(new IgnoreTestMutations());
             // Add a argument field with a require parameter
             var gql = new QueryRequest {
@@ -75,7 +75,7 @@ namespace EntityGraphQL.Tests
                     {"name", "Balance, Not Symmetry"},
                 }
             };
-            var results = new IgnoreTestSchema().QueryObject(gql, schemaProvider);
+            var results = schemaProvider.ExecuteQuery(gql, new IgnoreTestSchema(), null, null);
             dynamic data = results.Data["addAlbum"];
             Assert.Equal("Balance, Not Symmetry", data.name);
             Assert.Null(data.hiddenInputField); // not hidden from query
@@ -85,7 +85,7 @@ namespace EntityGraphQL.Tests
         [Fact]
         public void TestIgnoreAllInInput()
         {
-            var schemaProvider = SchemaBuilder.FromObject<IgnoreTestSchema>(false);
+            var schemaProvider = SchemaBuilder.FromObject<IgnoreTestSchema, object>(false);
             schemaProvider.AddMutationFrom(new IgnoreTestMutations());
             // Add a argument field with a require parameter
             var gql = new QueryRequest {
@@ -99,7 +99,7 @@ namespace EntityGraphQL.Tests
                     {"hiddenField", "yeh"},
                 }
             };
-            var results = new IgnoreTestSchema().QueryObject(gql, schemaProvider);
+            var results = schemaProvider.ExecuteQuery(gql, new IgnoreTestSchema(), null, null);
             var error = results.Errors.First();
             Assert.Equal("No argument 'hiddenField' found on field 'addAlbum'", error.Message);
         }
@@ -107,7 +107,7 @@ namespace EntityGraphQL.Tests
         [Fact]
         public void TestIgnoreAllInQuery()
         {
-            var schemaProvider = SchemaBuilder.FromObject<IgnoreTestSchema>(false);
+            var schemaProvider = SchemaBuilder.FromObject<IgnoreTestSchema, object>(false);
             // Add a argument field with a require parameter
             var gql = new QueryRequest {
                 Query = @"query Test {
@@ -117,7 +117,7 @@ namespace EntityGraphQL.Tests
 }",
                 Variables = new QueryVariables {}
             };
-            var results = new IgnoreTestSchema().QueryObject(gql, schemaProvider);
+            var results = schemaProvider.ExecuteQuery(gql, new IgnoreTestSchema(), null, null);
             var error = results.Errors.First();
             Assert.Equal("Field 'hiddenField' not found on current context 'Album'", error.Message);
         }
@@ -125,7 +125,7 @@ namespace EntityGraphQL.Tests
         [Fact]
         public void TestIgnoreWithSchema()
         {
-            var schemaProvider = SchemaBuilder.FromObject<IgnoreTestSchema>(false);
+            var schemaProvider = SchemaBuilder.FromObject<IgnoreTestSchema, object>(false);
             schemaProvider.AddMutationFrom(new IgnoreTestMutations());
             var schema = schemaProvider.GetGraphQLSchema();
             Assert.DoesNotContain("hiddenField", schema);
@@ -138,7 +138,7 @@ namespace EntityGraphQL.Tests
         [Fact]
         public void TestNotNullTypes()
         {
-            var schemaProvider = SchemaBuilder.FromObject<IgnoreTestSchema>(false);
+            var schemaProvider = SchemaBuilder.FromObject<IgnoreTestSchema, object>(false);
             schemaProvider.AddMutationFrom(new IgnoreTestMutations());
             var schema = schemaProvider.GetGraphQLSchema();
             // this exists as it is not null
@@ -147,7 +147,7 @@ namespace EntityGraphQL.Tests
         [Fact]
         public void TestNullableEnumInType()
         {
-            var schemaProvider = SchemaBuilder.FromObject<IgnoreTestSchema>(false);
+            var schemaProvider = SchemaBuilder.FromObject<IgnoreTestSchema, object>(false);
             var schema = schemaProvider.GetGraphQLSchema();
             // this exists as it is not null
             Assert.Contains("type Artist {\n\tid: Int!\n\ttype: ArtistType\n}", schema);
@@ -155,7 +155,7 @@ namespace EntityGraphQL.Tests
         [Fact]
         public void TestNotNullArgs()
         {
-            var schemaProvider = SchemaBuilder.FromObject<IgnoreTestSchema>(false);
+            var schemaProvider = SchemaBuilder.FromObject<IgnoreTestSchema, object>(false);
             schemaProvider.AddMutationFrom(new IgnoreTestMutations());
             var schema = schemaProvider.GetGraphQLSchema();
             // this exists as it is not null
@@ -164,7 +164,7 @@ namespace EntityGraphQL.Tests
         [Fact]
         public void TestNotNullEnumerableElementByDefault()
         {
-            var schemaProvider = SchemaBuilder.FromObject<IgnoreTestSchema>(false);
+            var schemaProvider = SchemaBuilder.FromObject<IgnoreTestSchema, object>(false);
             schemaProvider.AddMutationFrom(new IgnoreTestMutations());
             var schema = schemaProvider.GetGraphQLSchema();
             // this exists as it is not null
@@ -173,7 +173,7 @@ namespace EntityGraphQL.Tests
         [Fact]
         public void TestNullEnumerableElement()
         {
-            var schemaProvider = SchemaBuilder.FromObject<IgnoreTestSchema>(false);
+            var schemaProvider = SchemaBuilder.FromObject<IgnoreTestSchema, object>(false);
             schemaProvider.AddMutationFrom(new IgnoreTestMutations());
             var schema = schemaProvider.GetGraphQLSchema();
             // this exists as it is not null